(maint) Move ArtifactoryExtensions to a separate file
